% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_functions.R
\name{biber_spacy}
\alias{biber_spacy}
\title{This function takes a spacyr data object that has been
part-of-speech tagged and dependency parsed and
extracts counts of features that have been used in
Douglas Biber's research since the late 1980s.}
\usage{
biber_spacy(spacy_tks)
}
\arguments{
\item{spacy_tks}{A data.frame of tokens created by spacyr}
}
\value{
A data.frame of feature counts
}
\description{
Note that function sometimes relies on lists and a dictionary,
both of which are stored as R data.
At other times, the function identifies features
based on local cues, which are approximations.
}
\details{
The function returns a data.frame of normalized counts.
}
