% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_functions.R
\name{normalize_counts}
\alias{normalize_counts}
\title{Normalize to counts per 1,000 tokens}
\usage{
normalize_counts(counts)
}
\arguments{
\item{counts}{Data frame with numeric columns for counts of token, with one
row per document. Must include a \code{tot_counts} column with the total number
of tokens per document.}
}
\value{
\code{counts} data frame with counts normalized to rate per 1,000 tokens,
and \code{tot_counts} column removed
}
\description{
Normalize to counts per 1,000 tokens
}
\keyword{internal}
