% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_functions.R
\name{replace_nas}
\alias{replace_nas}
\title{Replace all NAs with 0}
\usage{
replace_nas(x)
}
\arguments{
\item{x}{Vector potentially containing NAs}
}
\description{
Replace all NAs with 0
}
\keyword{internal}
