% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{udpipe_samples}
\alias{udpipe_samples}
\alias{spacy_samples}
\title{Samples of parsed text}
\format{
An object of class \code{udpipe_connlu} of length 3.

An object of class \code{spacyr_parsed} (inherits from \code{data.frame}) with 1346 rows and 9 columns.
}
\usage{
udpipe_samples

spacy_samples
}
\description{
Examples of spaCy and udpipe tagging output from excerpts of several
public-domain texts. Can be passed to \code{biber()} to see examples of its
feature detection.
}
\details{
Texts consist of early paragraphs from several public-domain books
distributed by Project Gutenberg \url{https://gutenberg.org}. Document IDs are
the Project Gutenberg book numbers.

See \code{udpipe::udpipe_annotate()} and \code{spacyr::spacy_parse()} for
further details on the data format produced by each package.
}
\keyword{datasets}
