% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_functions.R
\name{biber_spacy}
\alias{biber_spacy}
\alias{biber_udpipe}
\title{Extract Biber features from a document parsed and annotated by spacyr or udpipe}
\usage{
biber_spacy(spacy_tks)

biber_udpipe(udpipe_tks)
}
\arguments{
\item{spacy_tks}{A \code{data.frame} of tokens created by \code{spacyr::spacy_parse()}}

\item{udpipe_tks}{A list or data frame of tokens parsed by
\code{udpipe::udpipe_annotate()}.}
}
\value{
A \code{data.frame} of features. Count features are normalized to the rate
per 1,000 tokens.
}
\description{
Takes data that has been part-of-speech tagged and dependency parsed and
extracts counts of features that have been used in Douglas Biber's research
since the late 1980s.
}
\details{
This relies on a feature dictionary (included as \code{dict}) and word lists
(\code{word_lists}) to match specific features; see their documentation and values
for details on the exact patterns and words matched by each. The function
identifies other features based on local cues, which are approximations.
}
\references{
Biber, Douglas (1988). \emph{Variation across Speech and Writing}. Cambridge
University Press.

Biber, Douglas (1995). \emph{Dimensions of Register Variation: A Cross-Linguistic
Comparison.} Cambridge University Press.
}
